postgis-jdbc-jts: Align Maven module name, Java module name, package name and directory name
