+++ postgis-jdbc-jts/src/main/java/net/postgis/jdbc/jts/JtsBinaryWriter.java
@@ -22,7 +22,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  * 
  */
-package net.postgis.jts;
+package net.postgis.jdbc.jts;
 
 import org.locationtech.jts.geom.CoordinateSequence;
 import org.locationtech.jts.geom.Geometry;
