+++ postgis-jdbc-jts/src/main/java/net/postgis/jdbc/jts/JtsGeometry.java
@@ -24,7 +24,7 @@
  * 
  */
 
-package net.postgis.jts;
+package net.postgis.jdbc.jts;
 
 import java.sql.SQLException;
