+++ postgis-jdbc-jts/src/test/resources-filtered/testng-it.xml
@@ -10,7 +10,7 @@
     <test name="JTS Parser Integration Tests">
         <classes>
             <class name="net.postgis.tools.testutils.TestContainerController"/>
-            <class name="net.postgis.jts.JtsParserTest"/>
+            <class name="net.postgis.jdbc.jts.JtsParserTest"/>
         </classes>
     </test>
