+++ postgis-jdbc-jts/src/test/resources/logback-test.xml
@@ -14,6 +14,6 @@
     <logger name="org.testcontainers" level="ERROR"/>
     <logger name="net.postgis" level="ERROR"/>
 
-    <logger name="net.postgis.jts" level="ERROR"/>
+    <logger name="net.postgis.jdbc.jts" level="ERROR"/>
 
 </configuration>
